requirements.txt
pandas
numpy
matplotlib
requests
python-binance
pyyaml
websockets
torch
tensorflow
